// Cxh_pulse_chirstmas_edition/frontend/src/types/sections.types.ts
export type PromptsTrendTableV1Props = {
  dateRange: { from: Date; to: Date };
  intent?: string;
  ward?: string;
  county?: string;
  onTitleClick?: () => void;
  isTitleClickable?: boolean;
  hideTitle?: boolean;
};